JOb_agent/requirements.txt
# Core Agent Framework
agno>=0.1.0

# Email Processing
imap-tools>=1.5.0
email-validator>=2.0.0
python-dotenv>=1.0.0

# AI/ML
openai>=1.0.0
google-generativeai>=0.3.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Utilities
pydantic>=2.0.0
python-dateutil>=2.8.0
schedule>=1.2.0
rich>=13.0.0

# Dashboard (Optional)
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
